requirements.txt
appdirs==1.4.3
cvxopt==1.1.9
cycler==0.10.0
functools32==3.2.3.post2
matplotlib==2.0.2
numpy==1.12.1
packaging==16.8
pandas==0.20.1
pandas-datareader==0.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.14.2
requests-file==1.4.2
requests-ftp==0.3.1
scikit-learn==0.18.1
scipy==0.19.0
seaborn==0.7.1
six==1.10.0
subprocess32==3.2.7
Jinja2==2.9.6
